use dbman to deploy the database

+++ docs/install/container/refresh_db.sh
@@ -36,10 +36,23 @@ PORT=5432
 DB=onix
 DBUSER=onix
 DBPWD=onix
-SPATH=${HOME}"/go/src/github.com/gatblau/onix/wapi/src/main/resources/db/install/4"
 
-# removes the container
-docker rm -f oxdb
+# check if dbman is running
+DBMAN_UP=$(docker inspect -f '{{.State.Running}}' dbman)
+# if dbman is not running
+if [ ! "$DBMAN_UP" = "true" ]; then
+  # launch dbman
+  docker run --name dbman -itd -p 8085:8085 -e OX_DBM_DB_HOST=oxdb "gatblau/dbman-snapshot"
+fi
+
+#check if the onix database is running
+OXDB_UP=$(docker inspect -f '{{.State.Running}}' oxdb)
+
+# if oxdb is not running
+if [ "$OXDB_UP" = "true" ]; then
+  # removes the container
+  docker rm -f oxdb
+fi
 
 # re-creates the container
 docker run --name oxdb -it -d -p 5432:5432 -e POSTGRESQL_ADMIN_PASSWORD=${PGPASSWORD} "centos/postgresql-12-centos7"
@@ -50,21 +63,25 @@ sleep 5
 # shows the logs
 docker logs oxdb
 
-# re-deploys the database
-psql -h ${HOST} -U postgres -c "CREATE DATABASE "${DB}";"
-psql -h ${HOST} -U postgres -c "CREATE USER "${DBUSER}" WITH PASSWORD '"${DBPWD}"';"
-psql -h ${HOST} -U postgres ${DB} -c "CREATE EXTENSION IF NOT EXISTS hstore;"
-psql -h ${HOST} -U postgres ${DB} -c "CREATE EXTENSION IF NOT EXISTS intarray;"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/tables.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/json.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/validation.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/set.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/get.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/delete.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/queries.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/tree.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/tags.sql"
-psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/keyman.sql"
+# deploy database using dbman
+curl -H "Content-Type: application/json" -X POST http://localhost:8085/deploy/0.0.4 2>&1
 
+# uncomment below to use psql instead of dbman
+# re-deploys the database
+#SPATH=${HOME}"/go/src/github.com/gatblau/onix/wapi/src/main/resources/db/install/4"
+#psql -h ${HOST} -U postgres -c "CREATE DATABASE "${DB}";"
+#psql -h ${HOST} -U postgres -c "CREATE USER "${DBUSER}" WITH PASSWORD '"${DBPWD}"';"
+#psql -h ${HOST} -U postgres ${DB} -c "CREATE EXTENSION IF NOT EXISTS hstore;"
+#psql -h ${HOST} -U postgres ${DB} -c "CREATE EXTENSION IF NOT EXISTS intarray;"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/tables.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/json.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/validation.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/set.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/get.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/delete.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/queries.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/tree.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/tags.sql"
+#psql -h ${HOST} -U ${DBUSER} ${DB} -f "${SPATH}/keyman.sql"
